+++ src-tauri/crates/tenant-service/src/inventory/dto.rs
@@ -1,4 +1,3 @@
-use sea_orm::entity::prelude::Decimal;
 use sea_orm::FromQueryResult;
 use serde::{Deserialize, Serialize};
 use specta::Type;
@@ -19,10 +18,10 @@ pub struct SelectInventory {
     pub id: String,
     pub name: String,
     pub created_at: String,
-    #[specta(type = f64)]
-    pub price: Decimal,
-    #[specta(type = f64)]
-    pub quantity: Decimal,
+    #[specta(type = f32)]
+    pub price: f32,
+    #[specta(type = f32)]
+    pub quantity: f32,
     pub transaction_type: String,
 }
 
@@ -30,8 +29,8 @@ pub struct SelectInventory {
 pub struct NewInventory {
     pub transaction_type: String,
     pub product_id: String,
-    #[specta(type = f64)]
-    pub quantity: Decimal,
+    #[specta(type = f32)]
+    pub quantity: f32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Type)]
@@ -39,8 +38,8 @@ pub struct Inventory {
     pub id: String,
     pub transaction_type: String,
     pub product_id: String,
-    #[specta(type = f64)]
-    pub quantity: Decimal,
+    #[specta(type = f32)]
+    pub quantity: f32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Type)]
